#80: 1) Move extensions to org.javamodularity.moduleplugin.extensions package
Potential BREAKING CHANGE for Kotlin DSL.
