+++ src/main/java/org/javamodularity/moduleplugin/ModuleSystemPlugin.java
@@ -6,6 +6,7 @@ import org.gradle.api.plugins.ExtensionContainer;
 import org.gradle.api.plugins.JavaPlugin;
 import org.javamodularity.moduleplugin.extensions.DefaultModularityExtension;
 import org.javamodularity.moduleplugin.extensions.ModularityExtension;
+import org.javamodularity.moduleplugin.extensions.PatchModuleExtension;
 import org.javamodularity.moduleplugin.tasks.*;
 
 public class ModuleSystemPlugin implements Plugin<Project> {
